+++ core/src/executive/executive.rs
@@ -357,10 +357,10 @@ impl<'a> CallCreateExecutive<'a> {
     }
 
     fn enact_result(
-        result: &vm::Result<FinalizationResult>, state: &mut State,
+        result: vm::Result<FinalizationResult>, state: &mut State,
         substate: &mut Substate, mut unconfirmed_substate: Substate,
         sender: &Address, storage_limit: &U256, is_bottom_ex: bool,
-    ) -> CollateralCheckResult
+    ) -> vm::Result<FinalizationResult>
     {
         substate.pop_callstack_contract(&mut unconfirmed_substate);
         match result {
@@ -384,16 +384,15 @@ impl<'a> CallCreateExecutive<'a> {
                 apply_state: false, ..
             }) => {
                 state.revert_to_checkpoint();
-                CollateralCheckResult::Valid
+                result
             }
             Err(vm::Error::Reentrancy) => {
                 assert!(unconfirmed_substate.reentrancy_encountered);
                 state.discard_checkpoint();
-                CollateralCheckResult::Valid
+                result
             }
-            Err(vm::Error::StateDbError(e)) => {
-                panic!("db error occurred during execution, {}", e);
-            }
+            // The whole epoch execution fails. No need to revert state.
+            Err(vm::Error::StateDbError(_)) => result,
             Ok(_) => {
                 let check_result = if is_bottom_ex {
                     state.check_collateral_for_storage_finally(
@@ -403,25 +402,29 @@ impl<'a> CallCreateExecutive<'a> {
                     )
                 } else {
                     state.checkout_ownership_changed(&mut unconfirmed_substate)
-                };
+                }?;
                 match check_result {
-                    Ok(CollateralCheckResult::ExceedStorageLimit {
-                        ..
-                    }) => {
+                    CollateralCheckResult::ExceedStorageLimit { .. } => {
                         state.revert_to_checkpoint();
+                        Err(vm::Error::ExceedStorageLimit)
                     }
-                    Ok(CollateralCheckResult::NotEnoughBalance { .. }) => {
+                    CollateralCheckResult::NotEnoughBalance {
+                        required,
+                        got,
+                    } => {
                         state.revert_to_checkpoint();
+                        Err(vm::Error::NotEnoughBalanceForStorage {
+                            required,
+                            got,
+                        })
                     }
-                    Ok(CollateralCheckResult::Valid) => {
+                    CollateralCheckResult::Valid => {
                         state.discard_checkpoint();
                         substate.accrue(unconfirmed_substate);
-                    }
-                    Err(_) => {
-                        panic!("db error occurred during execution");
+
+                        result
                     }
                 }
-                check_result.unwrap()
             }
         }
     }
@@ -721,27 +724,15 @@ impl<'a> CallCreateExecutive<'a> {
                     }
                 };
 
-                match Self::enact_result(
-                    &res,
+                Ok(Self::enact_result(
+                    res,
                     state,
                     substate,
                     unconfirmed_substate,
                     &sender,
                     &storage_limit,
                     self.depth == 0,
-                ) {
-                    CollateralCheckResult::Valid => Ok(res),
-                    CollateralCheckResult::ExceedStorageLimit { .. } => {
-                        Ok(Err(vm::Error::ExceedStorageLimit))
-                    }
-                    CollateralCheckResult::NotEnoughBalance {
-                        required,
-                        got,
-                    } => Ok(Err(vm::Error::NotEnoughBalanceForStorage {
-                        required,
-                        got,
-                    })),
-                }
+                ))
             }
 
             CallCreateExecutiveKind::ExecCreate(
@@ -831,27 +822,15 @@ impl<'a> CallCreateExecutive<'a> {
                     }
                 };
 
-                match Self::enact_result(
-                    &res,
+                Ok(Self::enact_result(
+                    res,
                     state,
                     substate,
                     unconfirmed_substate,
                     &sender,
                     &storage_limit,
                     self.depth == 0,
-                ) {
-                    CollateralCheckResult::Valid => Ok(res),
-                    CollateralCheckResult::ExceedStorageLimit { .. } => {
-                        Ok(Err(vm::Error::ExceedStorageLimit))
-                    }
-                    CollateralCheckResult::NotEnoughBalance {
-                        required,
-                        got,
-                    } => Ok(Err(vm::Error::NotEnoughBalanceForStorage {
-                        required,
-                        got,
-                    })),
-                }
+                ))
             }
 
             CallCreateExecutiveKind::ResumeCall(..)
@@ -924,27 +903,15 @@ impl<'a> CallCreateExecutive<'a> {
                     }
                 };
 
-                match Self::enact_result(
-                    &res,
+                Ok(Self::enact_result(
+                    res,
                     state,
                     substate,
                     unconfirmed_substate,
                     &sender,
                     &storage_limit,
                     self.depth == 0,
-                ) {
-                    CollateralCheckResult::Valid => Ok(res),
-                    CollateralCheckResult::ExceedStorageLimit { .. } => {
-                        Ok(Err(vm::Error::ExceedStorageLimit))
-                    }
-                    CollateralCheckResult::NotEnoughBalance {
-                        required,
-                        got,
-                    } => Ok(Err(vm::Error::NotEnoughBalanceForStorage {
-                        required,
-                        got,
-                    })),
-                }
+                ))
             }
             CallCreateExecutiveKind::ResumeCreate(..) => {
                 panic!("Resumable as create, but called resume_call")
@@ -1022,27 +989,15 @@ impl<'a> CallCreateExecutive<'a> {
                     }
                 };
 
-                match Self::enact_result(
-                    &res,
+                Ok(Self::enact_result(
+                    res,
                     state,
                     substate,
                     unconfirmed_substate,
                     &sender,
                     &storage_limit,
                     self.depth == 0,
-                ) {
-                    CollateralCheckResult::Valid => Ok(res),
-                    CollateralCheckResult::ExceedStorageLimit { .. } => {
-                        Ok(Err(vm::Error::ExceedStorageLimit))
-                    }
-                    CollateralCheckResult::NotEnoughBalance {
-                        required,
-                        got,
-                    } => Ok(Err(vm::Error::NotEnoughBalanceForStorage {
-                        required,
-                        got,
-                    })),
-                }
+                ))
             }
             CallCreateExecutiveKind::ResumeCall(..) => {
                 panic!("Resumable as call, but called resume_create")
@@ -1716,7 +1671,7 @@ impl<'a> Executive<'a> {
         //        )?;
 
         match result {
-            Err(vm::Error::StateDbError(e)) => bail!(e),
+            Err(vm::Error::StateDbError(e)) => bail!(e.0),
             Err(exception) => Ok(ExecutionOutcome::ExecutionErrorBumpNonce(
                 ExecutionError::VmError(exception),
                 Executed::execution_error_fully_charged(tx),