+++ core/src/vm/context.rs
@@ -27,7 +27,7 @@ use super::{
     return_data::ReturnData,
     spec::Spec,
 };
-use crate::bytes::Bytes;
+use crate::{bytes::Bytes, statedb};
 use cfx_types::{Address, H256, U256};
 use std::sync::Arc;
 
@@ -100,7 +100,7 @@ pub trait Context {
     fn create(
         &mut self, gas: &U256, value: &U256, code: &[u8],
         address: CreateContractAddress, trap: bool,
-    ) -> ::std::result::Result<ContractCreateResult, TrapKind>;
+    ) -> statedb::Result<::std::result::Result<ContractCreateResult, TrapKind>>;
 
     /// Message call.
     ///
@@ -111,7 +111,7 @@ pub trait Context {
         &mut self, gas: &U256, sender_address: &Address,
         receive_address: &Address, value: Option<U256>, data: &[u8],
         code_address: &Address, call_type: CallType, trap: bool,
-    ) -> ::std::result::Result<MessageCallResult, TrapKind>;
+    ) -> statedb::Result<::std::result::Result<MessageCallResult, TrapKind>>;
 
     /// Returns code at given address
     fn extcode(&self, address: &Address) -> Result<Option<Arc<Bytes>>>;