+++ core/src/vm/tests.rs
@@ -28,7 +28,7 @@ use super::{
     CreateContractAddress, Env, GasLeft, MessageCallResult, Result, ReturnData,
     Spec,
 };
-use crate::{bytes::Bytes, hash::keccak};
+use crate::{bytes::Bytes, hash::keccak, statedb};
 use cfx_types::{Address, H256, U256};
 
 pub struct MockLogEntry {
@@ -147,7 +147,7 @@ impl Context for MockContext {
     fn create(
         &mut self, gas: &U256, value: &U256, code: &[u8],
         address: CreateContractAddress, _trap: bool,
-    ) -> ::std::result::Result<ContractCreateResult, TrapKind>
+    ) -> statedb::Result<::std::result::Result<ContractCreateResult, TrapKind>>
     {
         self.calls.insert(MockCall {
             call_type: MockCallType::Create,
@@ -160,14 +160,14 @@ impl Context for MockContext {
             code_address: None,
         });
         // TODO: support traps in testing.
-        Ok(ContractCreateResult::Failed)
+        Ok(Ok(ContractCreateResult::Failed))
     }
 
     fn call(
         &mut self, gas: &U256, sender_address: &Address,
         receive_address: &Address, value: Option<U256>, data: &[u8],
         code_address: &Address, _call_type: CallType, _trap: bool,
-    ) -> ::std::result::Result<MessageCallResult, TrapKind>
+    ) -> statedb::Result<::std::result::Result<MessageCallResult, TrapKind>>
     {
         self.calls.insert(MockCall {
             call_type: MockCallType::Call,
@@ -180,7 +180,7 @@ impl Context for MockContext {
             code_address: Some(code_address.clone()),
         });
         // TODO: support traps in testing.
-        Ok(MessageCallResult::Success(*gas, ReturnData::empty()))
+        Ok(Ok(MessageCallResult::Success(*gas, ReturnData::empty())))
     }
 
     fn extcode(&self, address: &Address) -> Result<Option<Arc<Bytes>>> {